requirements.txt
asttokens==2.2.1
backcall==0.2.0
bcrypt==3.2.2
bottle==0.12.25
cffi==1.15.0
contourpy==1.0.7
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.38.0
ipython==8.5.0
jedi==0.18.2
kiwisolver==1.4.4
matplotlib==3.7.0
matplotlib-inline==0.1.6
networkx==3.0
numpy==1.24.2
packaging==23.0
pandas==1.4.2
paramiko==2.11.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyNaCl==1.5.0
pyparsing==3.0.9
pysftp==0.2.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.3.1
readline==6.2.4.1
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
transitions==0.9.0
wcwidth==0.2.6
